pyproject.toml
[tool.poetry]
name = "detect-russian-road-signs"
version = "0.1.0"
description = "Neural Network for detecting road signs of the Russian Federation"
authors = ["nhassl3 <matvey.kvasov.05@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
matplotlib = "^3.8.4"
pyyaml = "^6.0.1"
roboflow = "^1.1.28"
ultralytics = "8.2.12"
ipywidgets = "^8.1.2"
torch = { url = "https://download.pytorch.org/whl/cu121/torch-2.3.0%2Bcu121-cp310-cp310-win_amd64.whl#sha256=002027d18a9c054f08fe9cf7a729e041229e783e065a71349015dcccc9a7137e" }
torchaudio = { url = "https://download.pytorch.org/whl/cu121/torchaudio-2.3.0%2Bcu121-cp310-cp310-win_amd64.whl#sha256=969a49fb31afd3c0d1c4b01d74514d61c44e25d7de4dbc2c4fec317fe31dc702" }
torchvision = { url = "https://download.pytorch.org/whl/cu121/torchvision-0.18.0%2Bcu121-cp310-cp310-win_amd64.whl#sha256=4ab207a0f35c8c2a43da91f19ee9248520239633dc8e11a9e4a2e77b076bb9db" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
